move _sk_visual_block from sklearn.forecasting.compose._pipeline._Pipeline to _HeterogenousMetaEstimator
