+++ sktime/forecasting/compose/_pipeline.py
@@ -13,7 +13,6 @@ from sktime.forecasting.base._base import BaseForecaster
 from sktime.forecasting.base._delegate import _DelegatedForecaster
 from sktime.forecasting.base._fh import ForecastingHorizon
 from sktime.registry import is_scitype
-from sktime.utils._estimator_html_repr import _VisualBlock
 from sktime.utils.validation.series import check_series
 from sktime.utils.warnings import warn
 
@@ -275,18 +274,6 @@ class _Pipeline(_HeterogenousMetaEstimator, BaseForecaster):
         params3 = {"steps": [Detrender(), YfromX.create_test_instance()]}
 
         return [params1, params2, params3]
-
-    def _sk_visual_block_(self):
-        names, estimators = zip(*self._steps)
-
-        name_details = [str(est) for est in estimators]
-        return _VisualBlock(
-            "serial",
-            estimators,
-            names=names,
-            name_details=name_details,
-            dash_wrapped=False,
-        )
 
 
 # we ensure that internally we convert to pd.DataFrame for now
